requirements.txt
argcomplete==1.11.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==6.7
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
cycler==0.10.0
Flask==1.1.2
idna==2.7
importlib-metadata==1.6.1
itsdangerous==0.24
Jinja2==2.11.2
keyboard==0.13.5
kiwisolver==1.2.0
MarkupSafe==0.23
matplotlib==3.2.2
numpy==1.19.0
Pillow==7.2.0
plaid-python==5.0.0
pycparser==2.20
pynput==1.6.8
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
requests==2.20.0
scipy==1.5.2
selenium==3.141.0
six==1.15.0
sounddevice==0.4.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
urllib3==1.24.3
webdriver-manager==3.2.2
Werkzeug==1.0.1
wordcloud==1.8.0
zipp==3.1.0
